Crate tartan_bitfield

Define structures that wrap a number with accessors for certain bit ranges.

Example

bitfield! {
    // The structure will be a wrapper for a u32 value
    pub struct Example(u32) {
        // Accessors for field `a` will refer to the first four least significant
        // bytes of the wrapped value, bits 0, 1, 2, and 3. Note that like normal
        // Rust ranges, the end of the range is *exclusive*.
        //
        // The accessors will be public, and will take/return the four bits as a `u8`.
        [0..4] pub a: u8,

        // No accessors cover bits `4..6`. This is legal and can be used for reserved
        // bits. However, these bits will still affect equality for the struct as a
        // whole.

        // Accessors for field `b` will refer to the twelve bits starting at bit 6,
        // but they will not be public. They will take/return the 12 bits as a `u16`.
        [6..18] b: u16,

        // Note that this bit range overlaps with `b`. This is allowed.
        [16..20] pub c: u8,

        // Accessors for field `d` will take/return a boolean and refer to a single
        // bit. Note that the `bool` is implied and not specified after the name.
        [25] pub d,

        // This will cover the 6 most significant bits of the wrapped value, but
        // the getters will take/return a `SubFields` struct instead of `u8`. This is
        // useful for nested bitfields, but the `A as B` syntax works for any `B`
        // which implements `Into<A>` and `From<A>`.
        [26..32] pub e: u8 as SubFields,
    }
}

bitfield! {
    // All accessors on this field use booleans and refer to single bits
    pub struct SubFields(u8) {
        [0] pub zero,
        [1] pub one,
        [2] pub two,
        [3] pub three,
        [4] pub four,
        [5] pub five,
    }
}

// The struct can be initialized with a u32 value
let x = Example(0xfa84_9e1b);
assert_eq!(x.a(), 0xb_u8);
assert_eq!(x.b(), 0x278_u16);  // Private, but still can be used within the module
assert_eq!(x.c(), 0x4_u8);
assert_eq!(x.d(), true);
assert_eq!(x.e(), SubFields(0x3e_u8));
assert_eq!(x.e().zero(), false);
assert_eq!(x.e().five(), true);

// It can also be converted Into and From its underlying representation
let n: u32 = x.into();
let y: Example = n.into();
assert_eq!(n, 0xfa84_9e1b);
assert_eq!(x, y);

// Setters are all prefixed with `set_`. They have the same visibility as the getters.
let mut z = Example::default();
z.set_a(0xb);
z.set_b(0x278);
z.set_c(0x4);
z.set_d(true);
z.set_e(SubFields(0x3e));
assert_eq!(z, Example(0xfa04_9e0b));

// Reserved ranges influence equality, and they are all zero on `z`.
assert_ne!(z, x);

Macros

bitfield

Define a structure that wraps a number with accessors for certain bit ranges.

bitfield_accessors

Define getters and setters for certain bit ranges. The containing type must implement the Bitfield trait.

bitfield_without_debug

Same as the bitfield macro without a Debug implementation provided.

Traits

Bitfield

Marker trait implemented by types defined with the bitfield macro.

OverflowingShl

A type with an overflowing left shift operation. Also adds a saturating version.

OverflowingShr

A type with an overflowing right shift operation. Also adds a saturating version.

TruncateInto

A type whose values can be truncated into another type. This is more explicit than x as T.

Functions

get_bit

Get a boolean reflecting a single bit of the value.

get_bits

Extract a range of bits from the value, shifted so the first bit of the subset is the least significant bit of the result.

set_bit

Create a copy of the value with a single bit modified.

set_bits

Create a copy of the value with a subset of bits updated based on the passed value.