Function tartan_bitfield::get_bits

#[must_use]pub fn get_bits<T>(packed_val: T, lsb: u8, msb: u8) -> T where
    T: Default + OverflowingShl + OverflowingShr + Not<Output = T> + BitAnd<T, Output = T>, 

Extract a range of bits from the value, shifted so the first bit of the subset is the least significant bit of the result.

Bits are numbered starting with zero for the least significant bit. The range of bits in the result is lsb..msb, exclusive of msb.

assert_eq!(get_bits(0b1100_1110_u8, 3, 7), 0b1001);
assert_eq!(get_bits(0b1010_0101_u8, 6, 8), 0b10);