Function tartan_bitfield::set_bits

#[must_use]pub fn set_bits<T>(packed_val: T, lsb: u8, msb: u8, field_val: T) -> T where
    T: Default + Copy + OverflowingShl + Shl<u8, Output = T> + Not<Output = T> + BitAnd<T, Output = T> + BitOr<T, Output = T>, 

Create a copy of the value with a subset of bits updated based on the passed value.

Bits are numbered starting with zero for the least significant bit. The range of updated bits is lsb..msb, exclusive of msb. field_val is shifted left lsb bytes before being combined with packed_val.

assert_eq!(set_bits(0b0000_0000_u8, 6, 8, 0b11), 0b1100_0000);
assert_eq!(set_bits(0b1111_1111_u8, 1, 5, 0b0000), 0b1110_0001);
assert_eq!(set_bits(0b1010_0110_u8, 2, 6, 0b1110), 0b1011_1010);