Macro tartan_bitfield::bitfield

macro_rules! bitfield {
    [
        $( #[$meta:meta] )*
        $vis:vis struct $struct:ident($underlying_type:ty) {
            $($body:tt)*
        }
    ] => { ... };
}

Define a structure that wraps a number with accessors for certain bit ranges.

See the crate documentation for an example.

The structure will implement these traits, where T is the underlying type defined in parentheses immediately after the struct name.

Bitfield<T>

Debug

Default

Copy

Eq

Into<T>

From<T>