Macro tartan_bitfield::bitfield_without_debug

macro_rules! bitfield_without_debug {
    [
        $( #[$meta:meta] )*
        $vis:vis struct $struct:ident($underlying_type:ty) {
            $($body:tt)*
        }
    ] => { ... };
}

Same as the bitfield macro without a Debug implementation provided.

Since Debug is required by the Bitfield trait, the caller must provide their own implementation.