Trait tartan_bitfield::OverflowingShl

pub trait OverflowingShl where
    Self: Sized + Default, {
    fn overflowing_shl(self, n: u32) -> (Self, bool);

    fn saturating_shl(self, n: u32) -> Self { ... }
}

A type with an overflowing left shift operation. Also adds a saturating version.

All basic numeric types have this operation, but there is no corresponding trait in core::ops.

Required methods

fn overflowing_shl(self, n: u32) -> (Self, bool)

Shift the value left by n mod m bits, where m is the number of bits in the type. Return the shifted value along with a boolean indicating whether the shift count was wrapped.

Since this behavior is counterintuitive and practically useless, see saturating_shl for an alternative that behaves the way you probably expect.

Loading content...
Provided methods

fn saturating_shl(self, n: u32) -> Self

Shift the value left by n bits. If n is greater than or equal to the number of bits in this type, the result will be zero.

Loading content...