Trait tartan_bitfield::TruncateInto

pub trait TruncateInto<T> {
    fn truncate_into(self) -> T;
}

A type whose values can be truncated into another type. This is more explicit than x as T.

Required methods

fn truncate_into(self) -> T

Truncate the value to fit in the destination type