Macro tartan_bitfield::bitfield_accessors

macro_rules! bitfield_accessors {
    [
        $(
            $( #[$meta:meta] )*
            [ $lsb:literal $( .. $msb:literal )? ]
            $vis:vis $field:ident
            $( : $underlying_type:ty $( as $interface_type:ty )? )?
        ),*
        $(,)?
    ] => { ... };
    [
        @field getter
        $( #[$meta:meta] )*
        [ $bit:literal ]
        $vis:vis $field:ident
    ] => { ... };
    [
        @field setter
        $( #[$meta:meta] )*
        [ $bit:literal ]
        $vis:vis $field:ident
    ] => { ... };
    [
        @field $accessor_type:tt
        $( #[$meta:meta] )*
        [ $lsb:literal .. $msb:literal ]
        $vis:vis $field:ident
        : $field_type:ty
    ] => { ... };
    [
        @field getter
        $( #[$meta:meta] )*
        [ $lsb:literal .. $msb:literal ]
        $vis:vis $field:ident
        : $underlying_type:ty as $interface_type:ty
    ] => { ... };
    [
        @field setter
        $( #[$meta:meta] )*
        [ $lsb:literal .. $msb:literal ]
        $vis:vis $field:ident
        : $underlying_type:ty as $interface_type:ty
    ] => { ... };
}

Define getters and setters for certain bit ranges. The containing type must implement the Bitfield trait.

This is most commonly invoked by the bitfield macro, but it can be used on its own to define common accessors as part of a trait, as in the example below:

trait CommonFields: Bitfield<u32> {
    bitfield_accessors! {
        // Since these are part of a trait, the `pub` keyword must be omitted.
        [ 0.. 6] a: u8,
        [14]     b,
        [18..32] c: u16,
    }
}

bitfield! {
    struct SomeFields(u32) {
        [ 7] pub x,
        [16] pub y,
    }
}

bitfield! {
    struct OtherFields(u32) {
        [10] pub z,
        [12] pub q,
    }
}

impl CommonFields for SomeFields {}
impl CommonFields for OtherFields {}

let f = SomeFields(0xabcd_1234);
assert_eq!(f.a(), 0x34); // has accessors from CommonFields
assert_eq!(f.y(), true); // has accessors from SomeFields
//assert_eq!(f.z(), false); // COMPILE ERROR: no accessors from OtherFields

let g = OtherFields(0xabcd_1234);
assert_eq!(g.a(), 0x34); // has accessors from CommonFields
//assert_eq!(g.y(), true); // COMPILE ERROR: no accessors from SomeFields
assert_eq!(g.z(), false); // has accessors from OtherFields